requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
plotly>=5.0.0
streamlit>=1.25.0
openai>=1.3.0
python-dotenv>=0.19.0
jupyter>=1.0.0 